move SpeechWithoutPauses out of speech.py
